Split model into multiple files
